version with 10 slivers and 2 rows
synchronizing the animations required setting the duration of the
timeline. also broke the timeline setup and config out of it's own
function and into the main flow of the script.

after the timeline has played out setResetSlivers() is called which
reconfigures the starting points for the slivers. once that's complete
the timeline is invalidated and reset using the new start values.